+++ pd_book_tools/geometry/point.py
@@ -1,4 +1,10 @@
-from typing import cast, override
+import sys
+from typing import cast
+
+if sys.version_info >= (3, 12):
+    from typing import override
+else:
+    from typing_extensions import override  # pyright: ignore[reportUnreachable]
 
 from pydantic import GetCoreSchemaHandler
 from pydantic_core import CoreSchema, core_schema